Move common parts of commands to own class
