+++ Classes/Command/ReindexCommand.php
@@ -19,9 +19,9 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use TYPO3\CMS\Core\Core\Bootstrap;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Core\Utility\MathUtility;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Connection;
+use Kitodo\Dlf\Command\BaseCommand;
 use Kitodo\Dlf\Common\Document;
 
 /**
@@ -32,7 +32,7 @@ use Kitodo\Dlf\Common\Document;
  * @subpackage dlf
  * @access public
  */
-class ReindexCommand extends Command
+class ReindexCommand extends BaseCommand
 {
     /**
      * Configure the command by defining the name, options and arguments
@@ -94,10 +94,8 @@ class ReindexCommand extends Command
         $io = new SymfonyStyle($input, $output);
         $io->title($this->getDescription());
 
-        $startingPoint = 0;
-        if (MathUtility::canBeInterpretedAsInteger($input->getOption('pid'))) {
-            $startingPoint = MathUtility::forceIntegerInRange((int) $input->getOption('pid'), 0);
-        }
+        $startingPoint = getStartingPoint($input->getOption('pid'));
+
         if ($startingPoint == 0) {
             $io->error('ERROR: No valid PID (' . $startingPoint . ') given.');
             exit(1);
@@ -108,11 +106,8 @@ class ReindexCommand extends Command
             && !is_array($input->getOption('solr'))
         ) {
             $allSolrCores = $this->getSolrCores($startingPoint);
-            if (MathUtility::canBeInterpretedAsInteger($input->getOption('solr'))) {
-                $solrCoreUid = MathUtility::forceIntegerInRange((int) $input->getOption('solr'), 0);
-            } else {
-                $solrCoreUid = $allSolrCores[$input->getOption('solr')];
-            }
+            $solrCoreUid = $this->getSolrCoreUid($allSolrCores, $input->getOption('solr'));
+
             // Abort if solrCoreUid is empty or not in the array of allowed solr cores.
             if (empty($solrCoreUid) || !in_array($solrCoreUid, $allSolrCores)) {
                 $output_solrCores = [];
@@ -172,38 +167,6 @@ class ReindexCommand extends Command
         }
 
         $io->success('All done!');
-    }
-
-
-    /**
-     * Fetches all Solr cores on given page.
-     *
-     * @param int $pageId The UID of the Solr core or 0 to disable indexing
-     *
-     * @return array Array of valid Solr cores
-     */
-    protected function getSolrCores(int $pageId): array
-    {
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-            ->getQueryBuilderForTable('tx_dlf_solrcores');
-
-        $solrCores = [];
-        $result = $queryBuilder
-            ->select('uid', 'index_name')
-            ->from('tx_dlf_solrcores')
-            ->where(
-                $queryBuilder->expr()->eq(
-                    'pid',
-                    $queryBuilder->createNamedParameter((int) $pageId, Connection::PARAM_INT)
-                )
-            )
-            ->execute();
-
-        while ($record = $result->fetch()) {
-            $solrCores[$record['index_name']] = $record['uid'];
-        }
-
-        return $solrCores;
     }
 
     /**
